+++ assets/src/scripts/components/hydrograph/axes.js
@@ -1,5 +1,5 @@
 const { axisBottom, axisLeft } = require('d3-axis');
-const { timeDay, timeWeek, timeMonth } = require('d3-time');
+const { timeDay } = require('d3-time');
 const { timeFormat } = require('d3-time-format');
 const { createSelector } = require('reselect');
 
@@ -12,28 +12,38 @@ const { yLabelSelector } = require('./timeSeries');
 
 const { getCurrentDateRange, getCurrentParmCd } = require('../../selectors/timeSeriesSelector');
 
-const INTERVAL = {
-    P7D: timeDay,
-    P30D: timeWeek,
-    P1Y: timeMonth
-};
+const dateFormatter = timeFormat('%b %d');
 /**
  * Create an x and y axis for hydrograph
  * @param  {Object} xScale      D3 Scale object for the x-axis
  * @param  {Object} yScale      D3 Scale object for the y-axis
  * @param  {Number} yTickSize   Size of inner ticks for the y-axis
  * @param {String} parmCd - parameter code of timeseries to be shown on the graph.
- * @param {String} period - ISO duration for date range of the timeseries
+ * * @param {String} period - ISO duration for date range of the timeseries
  * @return {Object}             {xAxis, yAxis} - D3 Axis
  */
 export const createAxes = function({xScale, yScale}, yTickSize, parmCd, period) {
+    const formatter = function(d) {
+        if(d.getHours() === 12) {
+            return dateFormatter(d);
+        } else {
+            return null;
+        }
+    };
     // Create x-axis
     const xAxis = axisBottom()
         .scale(xScale)
         .ticks(INTERVAL[period] ? INTERVAL[period] : timeDay)
         .tickFormat(timeFormat('%b %d'))
         .tickSizeOuter(0);
-
+    // Create a second x-axis. This x-axis will overlay the first x-axis and have labels every 12 hours but no ticks.
+    // Labels not on the 12 hour points are removed with formatting.
+    const xAxisWithDateTimeLabels = axisBottom()
+        .scale(xScale)
+        .ticks(timeDay.hour)
+        .tickPadding(5)
+        .tickSize(0)
+        .tickFormat(formatter);
     // Create y-axis
     const tickDetails = getYTickDetails(yScale.domain(), parmCd);
     const yAxis = axisLeft()
@@ -44,7 +54,7 @@ export const createAxes = function({xScale, yScale}, yTickSize, parmCd, period) 
         .tickPadding(3)
         .tickSizeOuter(0);
 
-    return {xAxis, yAxis};
+    return {xAxis, xAxisWithDateTimeLabels, yAxis};
 };
 
 
@@ -72,7 +82,7 @@ export const axesSelector = createSelector(
 /**
  * Add x and y axes to the given svg node.
  */
-export const appendAxes = function(elem, {xAxis, yAxis, layout, yTitle}) {
+export const appendAxes = function(elem, {xAxis, xAxisWithDateTimeLabels, yAxis, layout, yTitle}) {
     const xLoc = {
         x: 0,
         y: layout.height - (layout.margin.top + layout.margin.bottom)
@@ -84,13 +94,19 @@ export const appendAxes = function(elem, {xAxis, yAxis, layout, yTitle}) {
     };
 
     // Remove existing axes before adding the new ones.
-    elem.selectAll('.x-axis, .y-axis').remove();
+    elem.selectAll('.x-axis, .x-axis-date-time-label, .y-axis').remove();
 
     // Add x-axis
     elem.append('g')
         .attr('class', 'x-axis')
         .attr('transform', `translate(${xLoc.x}, ${xLoc.y})`)
         .call(xAxis);
+
+    // Add the second x-axis, the one with the centered date/time labels, on top of the first x-axis
+    elem.append('g')
+        .attr('class', 'x-axis-date-time-label')
+        .attr('transform', `translate(${xLoc.x}, ${xLoc.y})`)
+        .call(xAxisWithDateTimeLabels);
 
     // Add y-axis and a text label
     elem.append('g')