+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Cursor location / tooltip defaults to last point in the time series.
+- Date labels moved to and centered in areas between midnight tick marks
 
 
 ## [0.7.0] - 2018-04-23