+++ CHANGELOG.md
@@ -12,6 +12,7 @@ the selected timeseries.
 
 ### Changed
 - Cursor location / tooltip defaults to last point in the time series.
+- Date labels moved to and centered in areas between midnight tick marks
 
 
 ## [0.7.0] - 2018-04-23